refactor: move record stream classes
