+++ eze/src/main/kotlin/org/camunda/community/eze/records/ProcessInstanceRecordStream.kt
@@ -5,39 +5,39 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package org.camunda.community.eze
+package org.camunda.community.eze.records
 
 import io.camunda.zeebe.protocol.record.Record
 import io.camunda.zeebe.protocol.record.value.ProcessInstanceRecordValue
 
 class ProcessInstanceRecordStream(private val records: Iterable<Record<ProcessInstanceRecordValue>>) :
-        Iterable<Record<ProcessInstanceRecordValue>> {
+    Iterable<Record<ProcessInstanceRecordValue>> {
 
-	override fun iterator(): Iterator<Record<ProcessInstanceRecordValue>> {
+    override fun iterator(): Iterator<Record<ProcessInstanceRecordValue>> {
         return records.iterator()
     }
 
-	fun withBpmnProcessId(bpmnProcessId: String): ProcessInstanceRecordStream {
+    fun withBpmnProcessId(bpmnProcessId: String): ProcessInstanceRecordStream {
         return ProcessInstanceRecordStream(filter { it.value.bpmnProcessId == bpmnProcessId })
     }
 
-	fun withProcessDefinitionKey(processDefinitionKey: Long): ProcessInstanceRecordStream {
+    fun withProcessDefinitionKey(processDefinitionKey: Long): ProcessInstanceRecordStream {
         return ProcessInstanceRecordStream(filter { it.value.processDefinitionKey == processDefinitionKey })
     }
 
-	fun withElementId(elementId: String): ProcessInstanceRecordStream {
+    fun withElementId(elementId: String): ProcessInstanceRecordStream {
         return ProcessInstanceRecordStream(filter { it.value.elementId == elementId })
     }
 
-	fun withFlowScopeKey(flowScopeKey: Long): ProcessInstanceRecordStream {
+    fun withFlowScopeKey(flowScopeKey: Long): ProcessInstanceRecordStream {
         return ProcessInstanceRecordStream(filter { it.value.flowScopeKey == flowScopeKey })
     }
 
-	fun withParentProcessInstanceKey(parentProcessInstanceKey: Long): ProcessInstanceRecordStream {
+    fun withParentProcessInstanceKey(parentProcessInstanceKey: Long): ProcessInstanceRecordStream {
         return ProcessInstanceRecordStream(filter { it.value.parentProcessInstanceKey == parentProcessInstanceKey })
     }
 
-	fun withParentElementInstanceKey(parentElementInstanceKey: Long): ProcessInstanceRecordStream {
+    fun withParentElementInstanceKey(parentElementInstanceKey: Long): ProcessInstanceRecordStream {
         return ProcessInstanceRecordStream(filter { it.value.parentElementInstanceKey == parentElementInstanceKey })
     }
 }
